rename module to resolveToValueExternal
